changed 32 bit float to 64 bit for better precision

+++ cmd/gpud/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/an0nymoos3/gpud/internals/gpupricing"
 )
 
-func getInput() float32 {
-	var usd float32
+func getInput() float64 {
+	var usd float64
 
 	fmt.Print("Please enter price in USD: ")
 	fmt.Scan(&usd)
@@ -17,7 +17,7 @@ func getInput() float32 {
 func main() {
 	usd_price := getInput()
 	gpu_price := gpupricing.GetPrice()
-	gpud := usd_price / float32(gpu_price)
+	gpud := usd_price / float64(gpu_price)
 
 	fmt.Printf("$%f USD = $%f GPUD \n", usd_price, gpud)
 }
